// frontend/src/shared/lib/api/model/crossReferenceResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * prot-cellar
 * OpenAPI spec version: 0.1.0
 */
import type { CrossReferenceResponseUrl } from './crossReferenceResponseUrl';

export interface CrossReferenceResponse {
  database: string;
  accession: string;
  curie: string;
  url?: CrossReferenceResponseUrl;
}